custom hooks for Home Component
