+++ src/containers/use-search.js
@@ -1,0 +1,7 @@
+import { useState } from 'react';
+
+export function useSearch() {
+  const [search, setSearch] = useState('');
+  const handleChange = ({ target }) => setSearch(target.value);
+  return { search, handleChange };
+}
